+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- `state` do not add dot to empty prefix at own metrics.
+
+### Added
+- Add compression for both `receiver` and `sender`.
+- `state` new metrics: **OutBytes**, **OutBpm**.
+
 ## [1.3.0] - 2021-06-03
 ### Added
 - **hostname** argument. Allows setting hostname instead of getting a real one (good for Docker).